+++ src/Symfony/Component/Mime/CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
  * Support `binary` as a `Content-Transfer-Encoding`
  * Add PGP/MIME signing and encryption support with the `PgpSigner` and `PgpEncrypter` classes
+ * Add `AbstractPart::setContentTypeParameter()`
 
 8.0
 ---